Removed silly reference to a talking document
The way the document read before my edit suggested that "this manual uses the former pronunciation" as if to imply that the manual actually speaks. Even in the event that a screen reader were reading the manual, the pronunciation of SQL would be up to the screen reader, not the document writers. The writers of the document may suggest that users use one pronunciation or another, it may also state whether or not a pronunciation is the correct pronunciation according to the subject creators, but this should generally be avoided in any instances where pronunciation is debated and the creator(s) are not available to provide and authoritative position.